feat: migrating checkbox
