+++ eslint.config.cjs
@@ -101,18 +101,4 @@ module.exports = [
 			parser: require('jsonc-eslint-parser'),
 		},
 	},
-	{
-		files: ['**/*.json'],
-		rules: {
-			'@nx/dependency-checks': [
-				'error',
-				{
-					ignoredFiles: ['{projectRoot}/eslint.config.{js,cjs,mjs}'],
-				},
-			],
-		},
-		languageOptions: {
-			parser: require('jsonc-eslint-parser'),
-		},
-	},
 ];
