+++ libs/brain/checkbox/README.md
@@ -1,0 +1,3 @@
+# @spartan-ng/brain/checkbox
+
+Secondary entry point of `@spartan-ng/brain`. It can be used by importing from `@spartan-ng/brain/checkbox`.
