+++ libs/brain/checkbox/src/lib/brn-checkbox-ng-model.spec.ts
@@ -1,6 +1,27 @@
 import { render, screen } from '@testing-library/angular';
 import userEvent from '@testing-library/user-event';
-import { BrnCheckboxNgModelSpecComponent } from './brn-checkbox-ng-model.component.spec';
+
+import { Component, input, model } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { BrnCheckboxComponent } from './brn-checkbox.component';
+
+@Component({
+	selector: 'brn-checkbox-ng-model',
+	standalone: true,
+	template: `
+		<!-- eslint-disable-next-line @angular-eslint/template/label-has-associated-control -->
+		<label>
+			Airplane mode is: {{ airplaneMode() ? 'on' : 'off' }}
+			<brn-checkbox [disabled]="disabled()" [(ngModel)]="airplaneMode"></brn-checkbox>
+		</label>
+	`,
+	imports: [BrnCheckboxComponent, FormsModule],
+})
+export class BrnCheckboxNgModelSpecComponent {
+	public readonly disabled = input(false);
+
+	public readonly airplaneMode = model(false);
+}
 
 describe('BrnCheckboxComponentNgModelIntegration', () => {
 	const setup = async (airplaneMode = false, disabled = false) => {
