+++ libs/brain/package.json
@@ -1,9 +1,12 @@
 {
 	"name": "@spartan-ng/brain",
-	"version": "0.0.1",
+	"version": "0.0.1-alpha.356",
 	"peerDependencies": {
-		"@angular/common": "^18.2.0",
-		"@angular/core": "^18.2.0"
+		"@angular/common": ">=18.0.0",
+		"@angular/core": ">=18.0.0",
+		"@angular/cdk": ">=18.0.5",
+		"@angular/forms": ">=18.0.5",
+		"@spartan-ng/ui-forms-brain": "0.0.1-alpha.356"
 	},
 	"sideEffects": false
 }
