+++ libs/brain/src/index.ts
@@ -1,0 +1,2 @@
+// this file can't be empty, otherwise the build will fail
+export default true;
