+++ libs/cli/src/generators/migrate-brain-imports/import-map.json
@@ -1,3 +1,5 @@
 {
-	"imports": {}
+	"imports": {
+		"@spartan-ng/ui-checkbox-brain": "@spartan-ng/brain/checkbox"
+	}
 }
