+++ libs/tools/package.json
@@ -11,6 +11,9 @@
 		"jsonc-eslint-parser": "^2.1.0",
 		"replace-json-property": "^1.9.0",
 		"nx": "20.1.1",
-		"tslib": "~2.7.0"
+		"tslib": "~2.7.0",
+		"@nx/angular": "20.1.1",
+		"@nx/workspace": "20.1.1",
+		"@spartan-ng/cli": "0.0.1-alpha.366"
 	}
 }
