--- libs/ui/checkbox/brain/README.md
@@ -1,7 +1,0 @@
-# ui-checkbox-brain
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test ui-checkbox-brain` to execute the unit tests.
