--- libs/ui/checkbox/brain/eslint.config.js
@@ -1,34 +1,0 @@
-const nx = require('@nx/eslint-plugin');
-const baseConfig = require('../../../../eslint.config.cjs');
-
-module.exports = [
-	...baseConfig,
-	...nx.configs['flat/angular'],
-	...nx.configs['flat/angular-template'],
-	{
-		files: ['**/*.ts'],
-		rules: {
-			'@angular-eslint/directive-selector': [
-				'error',
-				{
-					type: 'attribute',
-					prefix: 'brn',
-					style: 'camelCase',
-				},
-			],
-			'@angular-eslint/component-selector': [
-				'error',
-				{
-					type: 'element',
-					prefix: 'brn',
-					style: 'kebab-case',
-				},
-			],
-		},
-	},
-	{
-		files: ['**/*.html'],
-		// Override or add rules here
-		rules: {},
-	},
-];
