--- libs/ui/checkbox/brain/jest.config.ts
@@ -1,21 +1,0 @@
-export default {
-	displayName: 'ui-checkbox-brain',
-	preset: '../../../../jest.preset.cjs',
-	setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
-	transform: {
-		'^.+\\.(ts|mjs|js|html)$': [
-			'jest-preset-angular',
-			{
-				tsconfig: '<rootDir>/tsconfig.spec.json',
-				stringifyContentPathRegex: '\\.(html|svg)$',
-			},
-		],
-	},
-	transformIgnorePatterns: ['node_modules/(?!.*\\.mjs$)'],
-	testPathIgnorePatterns: ['<rootDir>/src/lib/brn-checkbox-ng-model.component.spec.ts'],
-	snapshotSerializers: [
-		'jest-preset-angular/build/serializers/no-ng-attributes',
-		'jest-preset-angular/build/serializers/ng-snapshot',
-		'jest-preset-angular/build/serializers/html-comment',
-	],
-};
