--- libs/ui/checkbox/brain/ng-package.json
@@ -1,7 +1,0 @@
-{
-	"$schema": "../../../../node_modules/ng-packagr/ng-package.schema.json",
-	"dest": "../../../../dist/libs/ui/checkbox/brain",
-	"lib": {
-		"entryFile": "src/index.ts"
-	}
-}
