--- libs/ui/checkbox/brain/project.json
@@ -1,43 +1,0 @@
-{
-	"name": "ui-checkbox-brain",
-	"$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-	"sourceRoot": "libs/ui/checkbox/brain/src",
-	"prefix": "spartan",
-	"projectType": "library",
-	"tags": ["scope:brain"],
-	"targets": {
-		"build": {
-			"executor": "@nx/angular:package",
-			"outputs": ["{workspaceRoot}/dist/{projectRoot}"],
-			"options": {
-				"project": "libs/ui/checkbox/brain/ng-package.json"
-			},
-			"configurations": {
-				"production": {
-					"tsConfig": "libs/ui/checkbox/brain/tsconfig.lib.prod.json"
-				},
-				"development": {
-					"tsConfig": "libs/ui/checkbox/brain/tsconfig.lib.json"
-				}
-			},
-			"defaultConfiguration": "production"
-		},
-		"test": {
-			"executor": "@nx/jest:jest",
-			"outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-			"options": {
-				"jestConfig": "libs/ui/checkbox/brain/jest.config.ts"
-			}
-		},
-		"lint": {
-			"executor": "@nx/eslint:lint",
-			"outputs": ["{options.outputFile}"]
-		},
-		"release": {
-			"executor": "@spartan-ng/tools:build-update-publish",
-			"options": {
-				"libName": "ui-checkbox-brain"
-			}
-		}
-	}
-}
