--- libs/ui/checkbox/brain/tsconfig.lib.json
@@ -1,12 +1,0 @@
-{
-	"extends": "./tsconfig.json",
-	"compilerOptions": {
-		"outDir": "../../../../dist/out-tsc",
-		"declaration": true,
-		"declarationMap": true,
-		"inlineSources": true,
-		"types": []
-	},
-	"exclude": ["src/**/*.spec.ts", "src/test-setup.ts", "jest.config.ts", "src/**/*.test.ts"],
-	"include": ["src/**/*.ts"]
-}
