--- libs/ui/checkbox/brain/tsconfig.lib.prod.json
@@ -1,9 +1,0 @@
-{
-	"extends": "./tsconfig.lib.json",
-	"compilerOptions": {
-		"declarationMap": false
-	},
-	"angularCompilerOptions": {
-		"compilationMode": "partial"
-	}
-}
