--- libs/ui/checkbox/brain/tsconfig.spec.json
@@ -1,11 +1,0 @@
-{
-	"extends": "./tsconfig.json",
-	"compilerOptions": {
-		"outDir": "../../../../dist/out-tsc",
-		"module": "commonjs",
-		"target": "es2016",
-		"types": ["jest", "node"]
-	},
-	"files": ["src/test-setup.ts"],
-	"include": ["jest.config.ts", "src/**/*.test.ts", "src/**/*.spec.ts", "src/**/*.d.ts"]
-}
